Change game id to int64. Fixes #396

+++ LegendaryClient/Windows/FactionsGameLobbyPage.xaml.cs
@@ -94,7 +94,7 @@ namespace LegendaryClient.Windows
 
                     HasConnectedToChat = true;
                     string obfuscatedName =
-                        Client.GetObfuscatedChatroomName(dto.Name.ToLower() + Convert.ToInt32(dto.Id),
+                        Client.GetObfuscatedChatroomName(dto.Name.ToLower() + Convert.ToInt64(dto.Id),
                             ChatPrefixes.Arranging_Practice);
                     string Jid = Client.GetChatroomJid(obfuscatedName, dto.RoomPassword, false);
                     newRoom = new MucManager(Client.XmppConnection);
